refactor(comment): emotion selector styles re-org
